fix: changed div to span, and moved BG-logo
